apps/public/views.py
from django.views.generic import TemplateView


class HomeView(TemplateView):
    template_name = "public/home.html"

    def get_context_data(self, **kwargs):
        ctx = super().get_context_data(**kwargs)
        ctx['page_title'] = 'Welcome to Kenda'
        return ctx