requirements-dev.txt
# Include all production requirements
-r requirements.txt

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Code quality
black==23.12.1
flake8==7.0.0
isort==5.13.2
pylint-django==2.5.5
mypy==1.8.0
django-stubs==4.2.7

# Pre-commit
pre-commit==3.6.0

# Debug tools
django-debug-toolbar==4.2.0
django-silk==5.1.0
ipython==8.19.0
ipdb==0.13.13